Move non-v1 tests up a level
