+++ api-server/stack-test-suite/tests/v1/mod.rs
@@ -22,6 +22,7 @@ mod chain_tip;
 mod transaction;
 mod transaction_merkle_path;
 
+use crate::spawn_webserver;
 use api_blockchain_scanner_lib::{
     blockchain_state::BlockchainState, sync::local_state::LocalBlockchainState,
 };
@@ -40,52 +41,6 @@ use rstest::rstest;
 use serde_json::json;
 use std::{net::TcpListener, sync::Arc};
 use test_utils::random::{make_seedable_rng, Rng, Seed};
-
-async fn spawn_webserver(url: &str) -> (tokio::task::JoinHandle<()>, reqwest::Response) {
-    let listener = TcpListener::bind("127.0.0.1:0").unwrap();
-    let addr = listener.local_addr().unwrap();
-
-    let task = tokio::spawn(async move {
-        let web_server_state = {
-            let chain_config = Arc::new(create_unit_test_config());
-            let storage = TransactionalApiServerInMemoryStorage::new(&chain_config);
-
-            ApiServerWebServerState {
-                db: Arc::new(storage),
-                chain_config: Arc::clone(&chain_config),
-            }
-        };
-
-        web_server(listener, web_server_state).await.unwrap();
-    });
-
-    // Given that the listener port is open, this will block until a response is made (by the web server, which takes the listener over)
-    let response = reqwest::get(format!("http://{}:{}{url}", addr.ip(), addr.port()))
-        .await
-        .unwrap();
-
-    (task, response)
-}
-
-#[tokio::test]
-async fn server_status() {
-    let (task, response) = spawn_webserver("/").await;
-
-    assert_eq!(response.status(), 200);
-    assert_eq!(response.text().await.unwrap(), r#"{"versions":["1.0.0"]}"#);
-
-    task.abort();
-}
-
-#[tokio::test]
-async fn bad_request() {
-    let (task, response) = spawn_webserver("/non-existent-url").await;
-
-    assert_eq!(response.status(), 400);
-    assert_eq!(response.text().await.unwrap(), r#"{"error":"Bad request"}"#);
-
-    task.abort();
-}
 
 #[tokio::test]
 async fn v1_chain_genesis() {
@@ -116,7 +71,9 @@ async fn v1_chain_genesis() {
         }
     });
 
-    // Given that the listener port is open, this will block until a response is made (by the web server, which takes the listener over)
+    // Given that the listener port is open, this will block until a
+    // response is made (by the web server, which takes the listener
+    // over)
     let response = reqwest::get(format!("http://{}:{}{url}", addr.ip(), addr.port()))
         .await
         .unwrap();
